Menus Connected
all Menus should be linked to each other, strange bug with hud, but i think it can be fix by linking it to something else bug will occur in all slides.

Goofy thing there the camera will switch views and will look out of the thetre instead of in, think is is also apart of the  player controller.